refactor: move stage model stuff to dedicated file
